Move getAgentInfo to agent package
